requirements.txt
Flask==2.2.3
transformers
Pillow
requests
torch
flask-cors
